requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
openai==1.88.0
faiss-cpu==1.11.0
pillow==11.2.1
pytesseract==0.3.13
beautifulsoup4==4.13.4
requests==2.32.3
python-dateutil==2.9.0.post0
Markdown==3.8
python-dotenv==1.1.0
tiktoken==0.9.0
sentence-transformers==4.1.0
torch==2.7.1